+++ spec/functional/consumer_group_spec.rb
@@ -113,6 +113,47 @@ describe "Consumer API", functional: true do
     expect(received_messages.map(&:value).map(&:to_i).sort).to match_array messages
   end
 
+  example "subscribing to multiple topics using regex and enable refreshing the topic list" do
+    topic_a = generate_topic_name
+    topic_b = generate_topic_name
+
+    messages_a = (1..500).to_a
+    messages_b = (501..1000).to_a
+    messages = messages_a + messages_b
+
+    producer = Kafka.new(kafka_brokers, client_id: "test").producer
+
+    messages_a.each { |i| producer.produce(i.to_s, topic: topic_a) }
+    producer.deliver_messages
+
+    group_id = "test#{rand(1000)}"
+
+    received_messages = []
+
+    kafka = Kafka.new(kafka_brokers, client_id: "test", logger: logger)
+    consumer = kafka.consumer(group_id: group_id, offset_retention_time: offset_retention_time, refresh_topic_interval: 1)
+    consumer.subscribe(/#{topic_a}|#{topic_b}/)
+
+    thread = Thread.new do
+      consumer.each_message do |message|
+        received_messages << message
+
+        if received_messages.count == messages.count
+          consumer.stop
+        end
+      end
+    end
+    thread.abort_on_exception = true
+
+    sleep 1
+    messages_b.each { |i| producer.produce(i.to_s, topic: topic_b) }
+    producer.deliver_messages
+
+    thread.join
+
+    expect(received_messages.map(&:value).map(&:to_i).sort).to match_array messages
+  end
+
   example "consuming messages from a topic that's being written to" do
     num_partitions = 3
     topic = create_random_topic(num_partitions: num_partitions)
@@ -392,6 +433,80 @@ describe "Consumer API", functional: true do
       consumer_1_thread && consumer_1_thread.kill
       consumer_2_thread && consumer_2_thread.kill
     end
+  end
+
+  example "consuming messages with a custom assignment strategy" do
+    num_partitions = 3
+    topic = create_random_topic(num_partitions: num_partitions)
+    messages = (1..100 * num_partitions).to_a
+
+    begin
+      kafka = Kafka.new(kafka_brokers, client_id: "test")
+      producer = kafka.producer
+
+      messages.each do |i|
+        producer.produce(i.to_s, topic: topic, partition: i % num_partitions)
+      end
+
+      producer.deliver_messages
+    end
+
+    group_id = "test#{rand(1000)}"
+
+    mutex = Mutex.new
+    received_messages = Hash.new {|h, k| h[k] = [] }
+
+    assignment_strategy_class = Class.new do
+      def initialize(weight)
+        @weight = weight
+      end
+
+      def user_data
+        @weight.to_s
+      end
+
+      def call(cluster:, members:, partitions:)
+        member_ids = members.flat_map {|id, metadata| [id] * metadata.user_data.to_i }
+        partitions_per_member = Hash.new {|h, k| h[k] = [] }
+        partitions.each_with_index do |partition, index|
+          partitions_per_member[member_ids[index % member_ids.count]] << partition
+        end
+
+        partitions_per_member
+      end
+    end
+
+    consumers = 2.times.map do |i|
+      assignment_strategy = assignment_strategy_class.new(i + 1)
+
+      kafka = Kafka.new(kafka_brokers, client_id: "test", logger: logger)
+      consumer = kafka.consumer(group_id: group_id, offset_retention_time: offset_retention_time, assignment_strategy: assignment_strategy)
+      consumer.subscribe(topic)
+      consumer
+    end
+
+    threads = consumers.map do |consumer|
+      t = Thread.new do
+        consumer.each_message do |message|
+          mutex.synchronize do
+            received_messages[consumer] << message
+
+            if received_messages.values.flatten.count == messages.count
+              consumers.each(&:stop)
+            end
+          end
+        end
+      end
+
+      t.abort_on_exception = true
+
+      t
+    end
+
+    threads.each(&:join)
+
+    expect(received_messages.values.flatten.map {|v| v.value.to_i }).to match_array messages
+    expect(received_messages.values.map(&:count)).to match_array [messages.count / 3, messages.count / 3 * 2]
   end
 
   def wait_until(timeout:)